refactor io.go
